Media Editor Experiment: Try adding a route-based version of the media editor
