+++ packages/media-editor/src/components/media-editor/style.scss
@@ -1,0 +1,89 @@
+@use "@wordpress/base-styles/colors" as *;
+@use "@wordpress/base-styles/mixins" as *;
+@use "@wordpress/base-styles/variables" as *;
+
+.media-editor {
+	display: flex;
+	flex: 1;
+	flex-direction: column;
+	height: 100%;
+	min-height: 0;
+
+	.media-editor__loading {
+		height: 100%;
+		display: flex;
+		align-items: center;
+		justify-content: center;
+	}
+
+	.interface-interface-skeleton {
+		position: relative;
+		top: auto;
+		right: auto;
+		bottom: auto;
+		left: auto;
+		height: 100%;
+		max-height: none;
+		flex: 1;
+		flex-direction: column;
+	}
+
+	.interface-interface-skeleton__editor {
+		min-width: 0;
+		flex: 1;
+	}
+
+	.interface-interface-skeleton__footer {
+		position: relative;
+		display: flex;
+	}
+
+	.interface-interface-skeleton.has-footer .interface-interface-skeleton__body {
+		padding-bottom: 0;
+	}
+
+	.media-editor__canvas {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		background: $gray-100;
+		overflow: auto;
+		height: 100%;
+
+		.media-editor-preview {
+			height: 100%;
+		}
+
+		img,
+		video {
+			max-width: 100%;
+			max-height: 100%;
+			object-fit: contain;
+		}
+	}
+
+	.media-editor__sidebar {
+		@include reset;
+	}
+
+	.media-editor__panel {
+		padding: $grid-unit-20;
+	}
+
+	.components-panel__header.media-editor__sidebar-header {
+		padding-left: 0;
+		padding-right: $grid-unit-10;
+
+		.components-button.has-icon {
+			padding: 0;
+
+			@include break-medium() {
+				display: flex;
+			}
+		}
+	}
+}
+
+.media-editor__snackbar {
+	@include snackbar-container();
+}
