+++ packages/media-editor/src/style.scss
@@ -1,6 +1,7 @@
 // Import component styles
 @use "./components/media-preview/style.scss" as *;
 @use "./image-editor/style.scss" as *;
+@use "./components/media-editor/style.scss" as *;
 @use "./components/media-editor-modal/style.scss" as *;
 @use "./components/media-editor-canvas/style.scss" as *;
 @use "./components/media-editor-toolbar/style.scss" as *;
