+++ routes/post-list/route.ts
@@ -10,6 +10,17 @@ import { notFound } from '@wordpress/route';
  */
 import { ensureView, viewToQuery } from './view-utils';
 
+function getEditLink( postType: string, postId: string ) {
+	if (
+		postType === 'attachment' &&
+		typeof window !== 'undefined' &&
+		( window as any ).__experimentalMediaEditor
+	) {
+		return `/media-editor/${ postId }`;
+	}
+	return `/types/${ postType }/edit/${ postId }`;
+}
+
 /**
  * Route configuration for post list.
  */
@@ -64,7 +75,7 @@ export const route = {
 				postType: params.type,
 				postId,
 				isPreview: true,
-				editLink: `/types/${ params.type }/edit/${ postId }`,
+				editLink: getEditLink( params.type, postId ),
 			};
 		}
 
@@ -83,7 +94,7 @@ export const route = {
 				postType: params.type,
 				postId,
 				isPreview: true,
-				editLink: `/types/${ params.type }/edit/${ postId }`,
+				editLink: getEditLink( params.type, postId ),
 			};
 		}
